pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "docs-capacitor"
version = "0.1.0"
description = "Configurable documentation freshness detection pipeline for Microsoft Docs"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Julia Kulla-Mader"},
]

dependencies = [
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "click>=8.0",
    "jinja2>=3.1.0",
    "jsonschema>=4.17.0",
]

[project.optional-dependencies]
llm = [
    "openai>=1.0.0",
]
learn-auth = [
    "azure-identity>=1.12.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
all = [
    "docs-capacitor[llm,learn-auth,dev]",
]

[project.scripts]
capacitor = "capacitor.cli:main"

[tool.setuptools.packages.find]
include = ["capacitor*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
